[xy] Copy pipeline config from template folder.

+++ mage_ai/data_preparation/models/pipeline.py
@@ -1,6 +1,7 @@
 from mage_ai.data_cleaner.shared.utils import clean_name
 from mage_ai.data_preparation.models.block import Block
 import os
+import shutil
 import yaml
 
 PIPELINES_FOLDER = 'pipelines'
@@ -33,11 +34,15 @@ class Pipeline:
         pipeline_path = os.path.join(repo_path, PIPELINES_FOLDER, uuid)
         if os.path.exists(pipeline_path):
             raise Exception(f'Pipeline {name} alredy exists.')
-        os.makedirs(pipeline_path)
-        with open(os.path.join(pipeline_path, '__init__.py'), 'w'):
-            pass
-        with open(os.path.join(pipeline_path, 'requirements.txt'), 'w'):
-            pass
+        # Copy pipeline files from template folder
+        template_path = os.path.join(
+            os.path.dirname(os.path.dirname(__file__)),
+            'templates/pipeline',
+        )
+        if not os.path.exists(template_path):
+            raise IOError('Could not find templates for pipeline.')
+        shutil.copytree(template_path, pipeline_path)
+        # Update metadata.yaml with pipeline config
         with open(os.path.join(pipeline_path, 'metadata.yaml'), 'w') as fp:
             yaml.dump(dict(name=name, uuid=uuid), fp)
         return Pipeline(uuid, repo_path)
